server/requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
mysql-connector-python==8.0.32
python-dotenv==1.0.0
py-bcrypt==0.4